fix(ui): arm history prefetch at the full 1200px lead
ClawSweeper caught that the wider observer rootMargin was inert: arming
still required scrollTop <= 300px, so loads kept starting at the old wall.
One CHAT_HISTORY_PREFETCH_EDGE_PX constant now owns both the intent-arming
gates and the observer margin. Regression test drives upward intent at
800px and observes the chat.history request (fails at the old 300px gate).
Splits the reply-navigation tests into their own suite to stay under the
test-file line cap.
